move multiple nodes test to separate file, delay bootstrap procedure
